Use a macro for HAL header redirection
Instead of using routing headers we can simplify the HAL layout by instead using a macro with the include statements, some files need renamed to remove redundant prefixes / postfixes
