+++ Marlin/src/HAL/HAL_DUE/HAL_timers_Due.cpp
@@ -32,7 +32,7 @@
 // Includes
 // --------------------------------------------------------------------------
 
-#include "../HAL.h"
+#include "HAL.h"
 
 #include "HAL_timers_Due.h"
