+++ Marlin/src/HAL/HAL_STM32F1/HAL.cpp
@@ -1,7 +1,10 @@
 /**
  * Marlin 3D Printer Firmware
+ *
  * Copyright (C) 2016 MarlinFirmware [https://github.com/MarlinFirmware/Marlin]
  * Copyright (c) 2016 Bob Cousins bobcousins42@googlemail.com
+ * Copyright (c) 2015-2016 Nico Tonnhofer wurstnase.reprap@gmail.com
+ * Copyright (c) 2017 Victor Perez
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,24 +18,23 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
  */
 
 /**
- * Description: HAL for Arduino Due and compatible (SAM3X8E)
- *
- * For ARDUINO_ARCH_SAM
+ * HAL for stm32duino.com based on Libmaple and compatible (STM32F1)
  */
 
-#ifdef ARDUINO_ARCH_SAM
+#ifdef __STM32F1__
 
 // --------------------------------------------------------------------------
 // Includes
 // --------------------------------------------------------------------------
 
-#include "../HAL.h"
+#include "HAL.h"
+#include <STM32ADC.h>
 
-#include <Wire.h>
-#include "usb/usb_task.h"
+//#include <Wire.h>
 
 // --------------------------------------------------------------------------
 // Externals
@@ -53,12 +55,66 @@
 // --------------------------------------------------------------------------
 // Public Variables
 // --------------------------------------------------------------------------
+USBSerial SerialUSB;
 
 uint16_t HAL_adc_result;
 
 // --------------------------------------------------------------------------
 // Private Variables
 // --------------------------------------------------------------------------
+STM32ADC adc(ADC1);
+
+uint8 adc_pins[] = {
+  #if HAS_TEMP_0
+    TEMP_0_PIN,
+  #endif
+  #if HAS_TEMP_1
+    TEMP_1_PIN
+  #endif
+  #if HAS_TEMP_2
+    TEMP_2_PIN,
+  #endif
+  #if HAS_TEMP_3
+    TEMP_3_PIN,
+  #endif
+  #if HAS_TEMP_4
+    TEMP_4_PIN,
+  #endif
+  #if HAS_TEMP_BED
+    TEMP_BED_PIN,
+  #endif
+  #if ENABLED(FILAMENT_WIDTH_SENSOR)
+    FILWIDTH_PIN,
+  #endif
+};
+
+enum TEMP_PINS : char {
+  #if HAS_TEMP_0
+    TEMP_0,
+  #endif
+  #if HAS_TEMP_1
+    TEMP_1,
+  #endif
+  #if HAS_TEMP_2
+    TEMP_2,
+  #endif
+  #if HAS_TEMP_3
+    TEMP_3,
+  #endif
+  #if HAS_TEMP_4
+    TEMP_4,
+  #endif
+  #if HAS_TEMP_BED
+    TEMP_BED,
+  #endif
+  #if ENABLED(FILAMENT_WIDTH_SENSOR)
+    FILWIDTH,
+  #endif
+    ADC_PIN_COUNT
+};
+
+uint16_t HAL_adc_results[ADC_PIN_COUNT];
+
 
 // --------------------------------------------------------------------------
 // Function prototypes
@@ -72,66 +128,101 @@ uint16_t HAL_adc_result;
 // Public functions
 // --------------------------------------------------------------------------
 
-// HAL initialization task
-void HAL_init(void) {
-  // Initialize the USB stack
-  #if ENABLED(SDSUPPORT)
-    OUT_WRITE(SDSS, HIGH);  // Try to set SDSS inactive before any other SPI users start up
-  #endif
-  usb_task_init();
-}
-
-// HAL idle task
-void HAL_idletask(void) {
-  // Perform USB stack housekeeping
-  usb_task_idle();
-}
-
-// Disable interrupts
+/* VGPV Done with defines
+// disable interrupts
 void cli(void) { noInterrupts(); }
 
-// Enable interrupts
+// enable interrupts
 void sei(void) { interrupts(); }
+*/
 
 void HAL_clear_reset_source(void) { }
 
-uint8_t HAL_get_reset_source(void) {
-  switch ((RSTC->RSTC_SR >> 8) & 0x07) {
-    case 0: return RST_POWER_ON;
-    case 1: return RST_BACKUP;
-    case 2: return RST_WATCHDOG;
-    case 3: return RST_SOFTWARE;
-    case 4: return RST_EXTERNAL;
-    default: return 0;
-  }
-}
+/**
+ * TODO: Check this and change or remove.
+ * currently returns 1 that's equal to poweron reset.
+ */
+uint8_t HAL_get_reset_source(void) { return 1; }
 
-void _delay_ms(const int delay_ms) {
-  // Todo: port for Due?
-  delay(delay_ms);
-}
+void _delay_ms(const int delay_ms) { delay(delay_ms); }
 
 extern "C" {
   extern unsigned int _ebss; // end of bss section
 }
 
-// Return free memory between end of heap (or end bss) and whatever is current
-int freeMemory() {
-  int free_memory, heap_end = (int)_sbrk(0);
-  return (int)&free_memory - (heap_end ? heap_end : (int)&_ebss);
+/**
+ * TODO: Change this to correct it for libmaple
+ */
+
+// return free memory between end of heap (or end bss) and whatever is current
+
+/*
+#include "wirish/syscalls.c"
+//extern caddr_t _sbrk(int incr);
+#ifndef CONFIG_HEAP_END
+extern char _lm_heap_end;
+#define CONFIG_HEAP_END ((caddr_t)&_lm_heap_end)
+#endif
+
+extern "C" {
+  static int freeMemory() {
+    char top = 't';
+    return &top - reinterpret_cast<char*>(sbrk(0));
+  }
+  int freeMemory() {
+    int free_memory;
+    int heap_end = (int)_sbrk(0);
+    free_memory = ((int)&free_memory) - ((int)heap_end);
+    return free_memory;
+  }
 }
+*/
 
 // --------------------------------------------------------------------------
 // ADC
 // --------------------------------------------------------------------------
+// Init the AD in continuous capture mode
+void HAL_adc_init(void) {
+  // configure the ADC
+  adc.calibrate();
+  adc.setSampleRate(ADC_SMPR_41_5); // ?
+  adc.setPins(adc_pins, ADC_PIN_COUNT);
+  adc.setDMA(HAL_adc_results, (uint16_t)ADC_PIN_COUNT, (uint32_t)(DMA_MINC_MODE | DMA_CIRC_MODE), (void (*)())NULL);
+  adc.setScanMode();
+  adc.setContinuous();
+  adc.startConversion();
+}
 
 void HAL_adc_start_conversion(const uint8_t adc_pin) {
-  HAL_adc_result = analogRead(adc_pin);
+  TEMP_PINS pin_index;
+  switch (adc_pin) {
+    #if HAS_TEMP_0
+      case TEMP_0_PIN: pin_index = TEMP_0; break;
+    #endif
+    #if HAS_TEMP_1
+      case TEMP_1_PIN: pin_index = TEMP_1; break;
+    #endif
+    #if HAS_TEMP_2
+      case TEMP_2_PIN: pin_index = TEMP_2; break;
+    #endif
+    #if HAS_TEMP_3
+      case TEMP_3_PIN: pin_index = TEMP_3; break;
+    #endif
+    #if HAS_TEMP_4
+      case TEMP_4_PIN: pin_index = TEMP_4; break;
+    #endif
+    #if HAS_TEMP_BED
+      case TEMP_BED_PIN: pin_index = TEMP_BED; break;
+    #endif
+    #if ENABLED(FILAMENT_WIDTH_SENSOR)
+      case FILWIDTH_PIN: pin_index = FILWIDTH; break;
+    #endif
+  }
+  HAL_adc_result = (HAL_adc_results[(int)pin_index] >> 2) & 0x3FF; // shift to get 10 bits only.
 }
 
 uint16_t HAL_adc_get_result(void) {
-  // nop
   return HAL_adc_result;
 }
 
-#endif // ARDUINO_ARCH_SAM
+#endif // __STM32F1__
