+++ Marlin/src/HAL/HAL_STM32F1/HAL_timers_Stm32f1.cpp
@@ -30,7 +30,7 @@
 // Includes
 // --------------------------------------------------------------------------
 
-#include "../HAL.h"
+#include "HAL.h"
 
 #include "HAL_timers_Stm32f1.h"
