+++ Marlin/src/HAL/HAL_STM32F7/EEPROM_Emul/eeprom_emul.h
@@ -9,7 +9,7 @@
   ******************************************************************************
   * @attention
   *
-  * <h2><center>&copy; Copyright © 2016 STMicroelectronics International N.V.
+  * <h2><center>&copy; Copyright Â© 2016 STMicroelectronics International N.V.
   * All rights reserved.</center></h2>
   *
   * Redistribution and use in source and binary forms, with or without
@@ -54,7 +54,7 @@
 // Includes
 // --------------------------------------------------------------------------
 #include "../../../inc/MarlinConfig.h"
-#include "../../HAL.h"
+#include "../HAL.h"
 
 /* Exported constants --------------------------------------------------------*/
 /* EEPROM emulation firmware error codes */
