+++ Marlin/src/HAL/HAL_STM32F7/EmulatedEeprom.cpp
@@ -30,7 +30,7 @@
 // Includes
 // --------------------------------------------------------------------------
 
-#include "../HAL.h"
+#include "HAL.h"
 #include "EEPROM_Emul/eeprom_emul.h"
