+++ Marlin/src/HAL/HAL_STM32F7/HAL_spi_STM32F7.cpp
@@ -36,7 +36,7 @@
 // Includes
 // --------------------------------------------------------------------------
 
-#include "../HAL.h"
+#include "HAL.h"
 #include "../HAL_SPI.h"
 #include "pins_arduino.h"
 #include "spi_pins.h"
