+++ Marlin/src/HAL/HAL_STM32F7/HAL_timers_STM32F7.cpp
@@ -26,7 +26,7 @@
 // Includes
 // --------------------------------------------------------------------------
 
-#include "../HAL.h"
+#include "HAL.h"
 
 #include "HAL_timers_STM32F7.h"
