+++ Marlin/src/HAL/HAL_STM32F7/SanityCheck.h
