+++ Marlin/src/HAL/HAL_STM32F7/pinsDebug.h
@@ -1,0 +1,1 @@
+#error Debug pins is not supported on this Platform!
