+++ Marlin/src/HAL/HAL_TEENSY35_36/HAL.cpp
@@ -25,7 +25,7 @@
 
 #if defined(__MK64FX512__) || defined(__MK66FX1M0__)
 
-#include "../HAL.h"
+#include "HAL.h"
 
 #include <Wire.h>
