+++ Marlin/src/HAL/HAL_TEENSY35_36/HAL_timers_Teensy.cpp
@@ -26,7 +26,7 @@
 
 #if defined(__MK64FX512__) || defined(__MK66FX1M0__)
 
-#include "../HAL.h"
+#include "HAL.h"
 #include "HAL_timers_Teensy.h"
